rril: remove wrong NULL check done for call barring queries.
BZ: 1343

Call Barring queries are supported by Android using MMI code like
*#35#  or *#351#. The RIL is returning an error as soon as the
password is NULL or the class is NULL because it's optional for
queries according to the 3GPP spec. This needs to be reviewed on
the RRIL side and we should handle thes cases where the password
is NULL and/or the class is NULL.

This patch review the different check done in the RRIL. In past
versions of Android, the parameters to the APIs in question were
never NULL, but the string was passed to RIL as "". Now the
parameters are passed as NULL so no need to error out in case of
NULL.
Signed-off-by: Guillaume Lucas <guillaumex.lucas@intel.com>

Change-Id: I3a05dd4c929ee3d2b9a86f4f28957d41bdfe38d8