rril: modem reset before end of core dump
BZ: 7952

When a modem reset occurs, the RRIL will start the core dump
retrieval. If the RRIL ask, for any reason, to do a reset of
the modem in the same time, this last one will occurs even
if the ongoing core dump retrieval is not yet completly done.

The issue occurs when the reset done by the RRIL occurs between
the time where the WatchdogProc thread as dectected the modem
reset and the time where this process really start the core
dump retrieval.

This patch fix the issue using the g_bSpoofCommands variable in
the TriggerRadioError to check if a modem reset is detected or
not.

Change-Id: I48dcc94d84c9a0bcd8f4d5fc95059c8472458227