[PORT FROM R1] fix: Wrong time displayed on phone
BZ: 8241

CTZDST will be reported only if the network daylight saving time is
received from network as part of the MM information message. Due to this,
RRIL was not reporting the time to the android. Fix is to schedule
the timer once the CTZV is received. This way, time information will be
reported always and also only once as expected.

Change-Id: Ida720b2cdd44b8184640ffcb1451ef24f42e7ec6
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/17211
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>