[PORT FROM R1] rapid_ril: Modified RIL_REQUEST_SET_LOCATION_UPDATES implementation
BZ: 10276

Google location service issues RIL_REQUEST_SET_LOCATION_UPDATES request
which disables the registration status URCs (+XREG). Because of this the
device is not getting the GPRS registration status and no further
registration status URCs. This change was done as part of BZ7549.

Now the usage of RIL_REQUEST_SET_LOCATION_UPDATES is clear, and for the
request RIL responds with success. This request is is followed by a query
to get registration status, and we will always use CREG=2;CREG?;CREG=0.

Change-Id: I4a9a34bbf838d676c931a500c35507758b7d1ad8
Orig-Change-Id: Id2e52c267993236abc658b28fe3284548138ee97
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Reviewed-on: http://android.intel.com:8080/19616
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>