[PORT FROM R1] rapid_ril: device shutdown procedure is not correct
 BZ 10352

The RRIL should monitor the system property sys.shutdown.requested intead of implement a HOOK to send AT+CFUN=0.
Thus, we are using the standard ShutdownThread code and the only modification is done in the RRIL.
The power down hook is also removed in this patch.
This patch depends on the 3 revert patches 19940, 19941 and 19942

Change-Id: I9a1edc20052b63cedd1a06fa86af18d1a61cbc9c
Orig-Change-Id: I7c4966a41839c5cfdb3afdf151c5aae2e44b0293
Signed-off-by: Frederic Predon <frederic.predon@intel.com>
Reviewed-on: http://android.intel.com:8080/20410
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>