rapid_ril: Modified rapid ril init and startup
BZ: 11894

Before the introduction of RESET_MGMT, rapid ril system manager
completes all its initialization and returns from InitializeSystem()
The initialization includes, initing and opening the ports, creating
the threads and mutexes, queues and modem initialization. But now the
opening of ports, and initialization of modem etc are not completed
before ril init is communicated to framework. This patch deferes this,
by introducing an event to wait till the init is completed.

InitializeSystem() function call will wait till the init is completed
when the modem is up and ContinueInit() is completed.

Change-Id: I34204796268c9cf27e9e8769d16d8d1c187ca169
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Reviewed-on: http://android.intel.com:8080/21148
Reviewed-by: Lucas, GuillaumeX <guillaumex.lucas@intel.com>
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Hubert, FabienX <fabienx.hubert@intel.com>
Reviewed-by: Lix, JackX W <jackx.w.lix@intel.com>
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>