[PORT FROM R2] rril: fix issue with null pointer access
BZ: 16331

Note: Dependency on the patch with change id:
I595e7d68a1b03f7d12f2de159dc2cbc8cb65f8e3

Incase of non-eng builds, ATCmdBuffer is freed as
soon as the command is written to port. After freeing,
same command buffer is passed to strlen function resulting
in crash.

Fix is to free the command buffer in non-engineering
builds if the number of retries is 0. Also pointer
is checked for NULL.

Change-Id: Ia959c7d6d851ea7e116dc5025f82b0fb55a6fbc2
Orig-Change-Id: If4efb5a80604384c5b44c34fc2f08760237feeed
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/27202
Reviewed-by: Delaude, EmmanuelX <emmanuelx.delaude@intel.com>
Reviewed-by: Lucas, GuillaumeX <guillaumex.lucas@intel.com>
Reviewed-by: Hubert, FabienX <fabienx.hubert@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>