rril: fix isse with lock/unlock requests
BZ: 16062

with RIL v6, AID has been added to all the lock/unlock
requests. RRIL validates the data size that is received
with the request before sending it to modem. Since it
is validating the size as per RIL v4, all the lock/unlock
requests are completed with error.

Fix is to check the required minimum required size rather than
checking the exact size.

Change-Id: Icc75f26b37fe86cf66bf01995408f7b6ee900e39
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/27209
Reviewed-by: Chen, Rong <rong.chen@intel.com>
Reviewed-by: Lucas, GuillaumeX <guillaumex.lucas@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Delaude, EmmanuelX <emmanuelx.delaude@intel.com>
Reviewed-by: Roques, CyrilX <cyrilx.roques@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>