[PORT FROM R2]rapid_ril: fix issue with USSD message coding
BZ: 23371

When the character set is not set by the TE using
AT+CSCS, default character set used is IRA. Since
not all the characters can be represented using
the IRA, some of the characters are not decoded
properly.

CUSD unsolicited responses are now enabled in the
channel 4. Since CSCS is channel specific setting,
character set is set to UCS2 for only channel 4.
Channel 4 is used for URCs.

Since there is a need also to support operator
names coded in UCS2 as well, ConvertUCS2ToUTF8 is
added to the utility functions.

Note: Character set for channel which is used for
sending USSD requests/responses is not set. Setting
the character set in other channels will require to
revisit all the commands which are dependent on
TE character set. This will be revisited based on
requirements.

Change-Id: Iab2aba2f1d429e6daa29188c84351fe6745a38f5
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/47683
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>