rril: fix missing 'goto Error'
BZ: 31380

If the call to the ioctl function for GSMIOC_ENABLE_NET failed,
there is no call to 'goto Error;'. This means that the RRIL will
continue the execution of the code as if all was good who is not
the case.

This patch adds the missing line 'goto Error;' in the code.
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>

Change-Id: I4478349a2975c8712ef4b8ef52ec91e9f555df46
Reviewed-on: http://android.intel.com:8080/42946
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Roulliere, Pierre <pierre.roulliere@intel.com>
Reviewed-by: Chen, Rong <rong.chen@intel.com>
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>