[PORT FROM R2] rril: retrieve channal data from CID in ParseSetupDataCall
BZ: 32738

In rare cases, we can have an issue occuring during the SETUP DATA
CALL. The ParseSetupDataCall function is executed without any error
but it's not the case of the following functions (ParseIpAddress and
ParseDns) who failed due to error to retrieve the channel data
associated to the CID.

The issue is due to the fact that the way used in the ParseSetupDataCall
function to retrieve the channel data structure is not the same the one
used in the other function. In ParseSetupDataCall the channel data is
retrieve according to the MUX channel and not to the CID.

This patch fix the ParseSetupDataCall function to retrieve the channel
data structure using the same way as the other function. Means by
the CID and not by the MUX channel.
Orig-Change-Id: Ic06cdf0fcb4153fb4116680c011ce7344f815855
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>

Change-Id: I7d3b2de5d9aa72b193a97d6afb692d360bad0c39
Reviewed-on: http://android.intel.com:8080/44390
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Roulliere, Pierre <pierre.roulliere@intel.com>
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Roques, CyrilX <cyrilx.roques@intel.com>
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>