rapid_ril : RADIO states are not managed properly in the RRIL
BZ: 33759

In the RRIL, the RADIO states are not managed properly. There is
a mix between the RADIO states and the SIM states.
This is not compatible with what the framework expects and leads
to problems when the device is encrypted.
With this patch, two states are managed by the RRIL, one for the
radio and one for the SIM. This is only when the framework queries
the RADIO state that we set the right value and send that back to the
framework.
Signed-off-by: Frederic Predon <frederic.predon@intel.com>

Change-Id: Ife93d05348f9458ab449b58876aa8cf9fd167cf2
Reviewed-on: http://android.intel.com:8080/47033
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>