rril: fix time out code for AT with no retries
BZ: 37762

If the PING sequence done when a time out occurred for an AT
command is successfull, the ParseResponse function is called
even if the AT command is not re-sent.

This patch fix the code who manage the timeout to exit the
function if an AT timed out when its number of retries is
set to 0.
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>

Change-Id: I6162118c8a36ad8d158bc97fd032f823909bc081