rril: fix time out code for AT with no retries
BZ: 37762

If the PING sequence done when a time out occurred for an AT
command is successfull, the ParseResponse function is called
even if the AT command is not re-sent.

This patch fix the code who manage the timeout to exit the
function if an AT timed out when its number of retries is
set to 0.
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>

Change-Id: I6162118c8a36ad8d158bc97fd032f823909bc081
Reviewed-on: http://android.intel.com:8080/49896
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Marmounier, EdwardX <edwardx.marmounier@intel.com>
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>