[PORT FROM R3] rapid_ril: Map CMS error codes to retry specific ril cause
BZ: 34365

If the send sms fails and if it can be retried, then
android expects the RIL_REQUEST_SEND_SMS to be completed
with error code RIL_E_SMS_SEND_FAIL_RETRY.

With this patch, few error codes are termed as
permanent failures for which retry should not be
done automatically. For all the other failure cases,
retry can be done.

Change-Id: Ie69dbf4de744eb257686eda2747aedf93795d687
Orig-Change-Id: Iee696369ee5464b4fca8e811ffbb267325e1b4be
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/52006
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>