rril: PIN2 related SIMIO handling using post command handler
BZ: 42615

Currently, PIN2 state is queried everytime when the
SIMIO request is issued with valid pin2.

Instead of quering everytime, PIN2 state cache is
added in this patch. Also the CoreSimIO is cleanedup
by moving the PIN2 provided SimIO requests to a new
function.

This patch is also modified to make use of the
post command handler.

Change-Id: I8f48e44885a7839c95fcce4988e2dc7eeef96796
Orig-Change-Id: I06dfc9185de4e9888e865940b8becb714d1ea5ef
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/62701
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Akue, LoicX <loicx.akue@intel.com>
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>