[PORT FROM R3] rril: return right error code for PLMN not supported by SIM
BZ: 48904

When a network selection manual is done to a PLMN not supported by
the SIM card, the RRIL will receive a CME ERROR code and returns
GENERIC_FAILURE to the framework. This is not the right value to
return according to the ril.h file.

This patch fix the RRIL code to return the ILLEGAL_SIM_OR_ME error
code for those cases.

Change-Id: I5505617c66b056d0df0dc37b62c848084879db35
Orig-Change-Id: Ie741e890f4c8f17c5a120ba5fa3b2f5c0c1f39e2
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>
Reviewed-on: http://android.intel.com:8080/60822
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Reviewed-by: Berat, FredericX <fredericx.berat@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>