rapid_ril: Add of new JB ril request RIL_REQUEST_VOICE_RADIO_TECH
BZ:50021

Add of new JB ril request RIL_REQUEST_VOICE_RADIO_TECH in rapid ril.
Align the comments in from of each ril request which id is >= 108 in rapid ril.
The implementation of the new request is not handle by this patch. The new request id has been
nevertheless added in preparation of its future implementation.

Change-Id: I6854a25375e7ae7531a1e4cd19fc21602072995d
Signed-off-by: alain kouassu <alain.kouassu@intel.com>
Reviewed-on: http://android.intel.com:8080/61805
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>