{PORT FROM R3] rril: avoid overwritting of IPV4 address in parsing of +CGPADDR
BZ: 49522

According to the 3GPP, the AT+CGPADDR response format is:

  AT+CGPADDR <cid> [,<PDP_addr_1>[,<PDP_addr_2>]]

Where <PDP_addr_1> is the IPV4 address and <PDP_addr_2> the IPV6
address.

In our case, the modem can return an IPV4V6 address for <PDP_addr_2>.
In this case, the RRIL will use this <PDP_addr_2> to retrieve the IPV4
address and doesn't use the <PDP_addr_1>. As there is an issue in the
IPV4 part of the IPV4V6 address returned by the modem, the IPV4 address
retrieved by the RRIL is not the correct one.

This patch fixes the RRIL code to always use the <PDP_addr_1> parameter
for the IPV4 address.

Change-Id: Ide80c07c3ed8dcf81035dd1784f6b72d6c90da2a
Orig-Change-Id: Ica629d21ed8513ccc3bf0b0481c892e4bf33a753
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>
Reviewed-on: http://android.intel.com:8080/60823
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Reviewed-by: Berat, FredericX <fredericx.berat@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>