[PORT FROM R3] rril: fix issue with USSD message conversion
BZ: 50530

As per 3GPP TS 23.038 section 5, if the DCS is 17
message is UCS2 coded and the message is preceded
by language indication.

This is not taken care when doing the conversion
from UCS2 to UTF8 resulting in wrong character shown
to user.

This patch fixes the issue by skipping the language
code before doing UCS2 to UTF8 conversion.

Change-Id: Ifbb451137087ea4b5522cd449f59ce39042a0749
Orig-Change-Id: Idc545e26f21e6505fcc18805c3bddfe74d3eb9d8
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Zhu Jian <jian.a.zhu@intel.com>
Reviewed-on: http://android.intel.com:8080/61771
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>