[PORT FROM DSDS] RRIL: review the shutdown sequence with multiple RRIL
BZ: 54339

When we use multiple RRILs, there is only one RRIL who will receive
the RADIO_POWER OFF request to perform the shutdown of the modem.
The others instances of the RRIL will be inform of the shutdown by
the STMD via the PLATFORM_SHUTDOWN event.

Actually the RRIL code do nothing when it receives this event. So
in case of multiple RRILs only the one who perfoms the shutdown of
the modem is correctly ended.

This patch fix the issue by implementing the necessary action to
perfom for the PLATFORM_SHUTDOWN event if the RRIL is not the one
who performs the shutdown. It fix also the ending of the threads
present in the RRIL to have a clean shutdown of all RRILs that could
be present in the system.

Change-Id: I2251f9daf8db2655b263571e89afb5ddfad03936
Orig-Change-Id: Ib6491565c0dc57fced42d23b46c2ee1ed7132a44
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>
Reviewed-on: http://android.intel.com:8080/65657
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>