[PORT FROM DSDS] rril: use blocking mode for the opening of the gsmtty ports
BZ: 54971

There is some rare cases where the modem doesn't respond to the first
SABM (open) request. In our platform the MUX automatically resend the
SABM request after 2,54 seconds if the modem doesn't send the UA
(acknowledge of SABM). In the RRIL when we do a write, if the DLCi is
still in the opening state we will receive the EAGAIN error. In this
case, the RRIL wait 100 ms and re-do the write. This timing is not
sufficient according to the timer use in the MUX. The RRIL doesn't let
enough time to the MUX to resend the SABM request.

This patch force the RRIL to open the gsmtty ports in blocking mode in
the goal to wait the modem respond before to perform the write
operation.

Change-Id: I589fa8a4978f76ec2cc3a0a38d5492af35bcfd7b
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>
Orig-Change-Id: If125b9248846412e68ded66f494b62e72a11718e
Reviewed-on: http://android.intel.com:8080/67507
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>