[PORT FROM R3] rril: improve the read thread blocking mechanism
BZ: 55943

If the BlockReadThread() function is called when the read thread
is blocked in the WaitForAvailableData() funtion, the thread will
not be really blocked. When the WaitForAvailableData() ends, the
thread reads the data and calls the parser.

This patch fix the RRIL code to avoid any call to the parsing
function even if the BlockReadThread() function was called after
the execution of the WaitForAvailableData(). When the read thread
is in blocking state, any data received during this state will be
discarded and not taken into account.

Change-Id: Ifa53b3062dca370ab477f8eb2783c9b0ac710ed4
Orig-Change-Id: Ib8be3e8b6ed4fa11309cc253868d2603cb76295c
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>
Reviewed-on: http://android.intel.com:8080/67119
Reviewed-by: Louis, NicolasX <nicolasx.louis@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>