rril: Ignore ME PDN ACT received with invalid CID
BZ: 60117

Some conformance test cases request to open the
PDP context manually. If the context is opened by
sending context activation commands directly to modem,
then rapid ril won't be aware of this. So, when a
ME PDN ACT is received with a context ID not initiated
by the rapid ril, this is treated as error resulting
in channel being blocked with the procoessing of
ME PDN ACT forever.

With this patch, external PDP context information
will be just ignored.

Commit: fd5e7c0c12e1b7d88832f4b84d0742a526452b1f
added the "goto Error;" due to which this issue
is seen.

Note: For LTE, this needs to be addressed in a separate
patch.

Change-Id: I0fb4b01a63ca1052b978a713952eab5367ac40fa
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/68990
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Delaude, EmmanuelX <emmanuelx.delaude@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>