rril: remove SetAlwaysParse for defining context
BZ: 63314

Currrently, AlwaysParse is set as true for
defining the context. This will result in
ParseSetupDataCall being called. Since,
ParseSetupDataCall returns RRIL_RESULT_OK, this
is set as the result code resulting in over
riding the actual result. Due to this CGACT is
issued even when CGDCONT has failed.

With the introduction of PostCommandHandler's,
SetAlwaysParse is not needed any more.

Remove the SetAlwaysParse issued for defining a
context.

Change-Id: I9df42ce6acd30d9e289b63fc8b94a40c445cd54f
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/71516
Reviewed-by: Goubert, TonyX <tonyx.goubert@intel.com>
Reviewed-by: Roques, CyrilX <cyrilx.roques@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>