[PORT FROM MAIN] Rapid_RIL: Integrating MMGR
BZ: 52800

MMGR (successor of STMD) provides anymore only one socket  to its clients.
This socket is accessible to radio group members and system user only.
The goal of this socket is to inform all clients of modem status and provide
a way for all its clients to ask for special operations.

All clients have to upgrade their code to handle the new interface:
* include the new MMGR header file: mmgr.h
* handle new information and notifications messages sent by MMGR
* handle new client request to MMGR.

As the resource get/release mechanism implies that RIL should be able
to handle a modem shutdown during flightmode, RRIL init is also reviewed.

Change-Id: I634460b4a650aed62883c2051be12fc1af8c863a
Orig-Change-Id: Ie8f3812de9231fd785cb960ea16d66b0479f048d
Signed-off-by: Frederic Berat <fredericx.berat@intel.com>
Reviewed-on: http://android.intel.com:8080/80232
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>