[PORT FROM MAIN] rril: Clear receive queue before sending PING command
BZ: 67926

Upon timeout of non-abortable AT command,
channel will be closed and reopened. After
reopening, PING command will be sent to modem
to check whether the modem is alive or dead.
This works fine if the timeout is due to
no response for an AT command from modem.
But, if the timeout is due to a garbage
response, this doesn't work.

Garbage response is not flushed after
reopening the channel and before sending
a PING command. Due to this, response for
PING command is not processed and gets
timed out after 3seconds.

The patch flushes the response after
reopening the channel and before sending
the PING command.

Change-Id: Id66be40be84166096887ee0f1b8a4f64cbd727a4
Orig-Change-Id: I63e707eac927518f3fcc4992da1edc12a9158eca
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/78796
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>