[PORT FROM DSDS] rril: delete the node upon dequeue
BZ: 78948

After dequeuing the node from the list, node
is not freed resulting in memory leak. Also, once
the object is found, control should exit the
for loop which is not done.

Fix the issue by deleting the dequeued node
and also break from the for loop once the
object is found.

Change-Id: Iaa12880cb854652d0759d411d627902261809797
Orig-Change-Id: I7ec2e7b7619cfc3afe613753502e248dc2d0fb54
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/86004
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Butscher, MorganeX <morganex.butscher@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>