[PORT FROM DSDS] rril: Fix system clock issue on airplane mode on/off
BZ: 84053

In the product builds, system clock time jumps
when airplane mode is switched on, sleep, wakeup
and then airplane mode switched off.

m_iFD is not initialized during object creation.
So, m_iFD gets initialized to 0. Closing of m_iFD(0)
ends up in closing the file descriptor which is used
in some other place. In this case, fd(0) is used
in rild for getting the elapsedRealTime.

When RILD call elapsedRealTime again, it will still use
fd(0) to get the local time which is not valid because
of rapid ril closing the fd(0).

So, wrong code branch gets called resulting in different
value to sent to framework causing time gaps.

Fix the issue by initializing m_iFd to -1 in the
constructor and also remove the CloseRepositoryFile
done in destructor as the close is already handled in
Read function.

Change-Id: Ic345c425dd3fdbd8fdac01e4fbd89949c7f40b50
Signed-off-by: Srivathsan Soundararajan <srivathsan.soundararajan@intel.com>
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/90028
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Roques, CyrilX <cyrilx.roques@intel.com>
Reviewed-by: Butscher, MorganeX <morganex.butscher@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>