[PORT FROM R4.1] rril: send init commands based on feature support
BZ: 79396

Currently, modem initialization commands sent
are not based on the modem and the device
features. Due to this, user would be able to
receiving incoming calls even when the device
supports only data. Even though device doesn't
have phone application to show the incoming
call screen but there are other issues seen
if the incoming call notification is forwarded
to android applications.

Following changes are done to address this issue:
- Modem and feature specific initialization commands
are removed from the basic initialization commands.
- Modem and feature specific Initialization commands
is provided by the modem specific object.

Change-Id: I65f37a9c7f1d3fe555a2e1ebc7ad610d44c96040
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/91879
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Butscher, MorganeX <morganex.butscher@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>