[PORT FROM R4.1] rril: Disable STK based on StkCapable flag
BZ: 81971

Currently, STK is enabled by default.
Even if the STK is disabled on the AP
side, STK is enabled on the modem side.
Even when the device doesn't support STK,
terminal profile is sent by modem to UICC.
As a result of this, device approval fails
and also this results in EAP-SIM not working.

Following changes are done to address this issue:
- StkCapable configuration flag is added
- If this flag is set to 0, then the AP
sends the command to fetch the proactive command
and disable the URC.

Change-Id: Icdb66d1cb088f90ed095b3121683c872475c14ba
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/91880
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Butscher, MorganeX <morganex.butscher@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>