rril: Manage PIN caching in flight mode - Modem off
BZ: 79165

Modem can be switched off in flight mode by
setting EnableModemOffInFlightMode to 1 in
/etc/rril/repository.txt. Currently, user
has to reenter the PIN on deactivation of
flight modem.

With this patch, if the PIN is required on
flight mode deactivation, rapid ril will
silently enter the cached pin if the cached
pin is for the inserted UICC and if the retry
count is > 1.

Change-Id: I34d938c5ff7c9e4473349f1a67a24cd916d053c3
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/91199
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Roques, CyrilX <cyrilx.roques@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Marmounier, Edward <edward.marmounier@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Butscher, MorganeX <morganex.butscher@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>