rril: cleanup channel info on modem reset cases
BZ: 72205

Currently, rapid ril is exiting on MODEM_DOWN,
COLD_RESET and MODEM_SHUTDOWN. When the user
deactivates the airplane mode with the flag
EnableModemOffInFlightMode set to 1, there is
toggling of airplane modem information in the
shutdown dialog. This is because of the rapid
ril exiting on modem down events.

Framework sets the radio state to RADIO_UNAVAILABLE
on rild disconnected. Toggling is because of this
state change.

Following changes done to address the issue
- Instead of exiting on modem down events, rapid ril
cleans up the channel information and sets the
initialization state to false.
- If the radio on/off request is ongoing, then the
RIL requests are completed with RADIO_NOT_AVAILABLE
This is also needed to avoid airplane mode information
toggling in shutdown dialog.
- 15seconds timeout is introduced for modem powered
up and basic initialization completion. If no signal
is received within the timeout, then RADIO_REQUEST
will be completed with error. Inorder to force the
framework to retry RADIO_POWER, radio state changed
will be changed to RADIO_UNAVAILABLE and then
RADIO_OFF. TODO: Timer value needs to be revisited.

- Also, with this patch retry done on port opening
failure is removed. Port opening done after MODEM_UP.
Once MODEM_UP is received, opening of ports should
always succeed.

Change-Id: I1ec336e4f57520e38d92a56dc3fa14a5fe5685b9
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/91202
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Roques, CyrilX <cyrilx.roques@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Berat, FredericX <fredericx.berat@intel.com>
Reviewed-by: Butscher, MorganeX <morganex.butscher@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>