rril: fix possible out of bound issue in encrypt function
BZ: 89370

If the nInputLen parameter of the encrypt function is greater than
BUF_LEN (9) we will try to access to buf[-1] which is an out of bound
error. According to this, the call to the encrypt function is not the
good one because we can have MAX_PIN_SIZE (10) for the nInputLen.

This patch fixes the code to avoid the out of bound issue. It also
removes the magic value of the encrypt() function using constant values.

Change-Id: I9ddc50e7e696f03ac9e27ff0002f736de82e125d
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>
Reviewed-on: http://android.intel.com:8080/94293
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Robert, Denis <denis.robert@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>