rril: update the channel use for SWITCH_WAITING_OR_HOLDING_AND_ACTIVE
BZ: 91404

In some case, we can have a hangup of a call performed before the
end of the switch from holding to active state operation. This is
due to the fact that those operations are not performed on the same
RRIL channel.

This patch update the RRIL code to perform the switch on the same
channel than the hangup and avoid to have one of those operation
did before the end of the other one.

Change-Id: I99620feb3ea505dda4d3aa2f3ec523459a1f2a7d
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>
Reviewed-on: http://android.intel.com:8080/96624
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>