[PORT FROM MAIN] rapid_ril: Manage access to ports though a mutex
BZ: 92208

When the device is in Flight mode, there could be other
mmgr client wishing to use modem. One such client is CWS.
In a particular use case where the CWS client tries to
acquire and release modem resource, there is possibility
that one thread in rril wants to open the gsmtty and use
while another thread tries to close the ports.

This patch controls the access to ports opening and closing
under a mutex to avoid this race condition.

Change-Id: Ia4f6f74380905ef2790b7cf2d2175f10183cc76f
Orig-Change-Id: Idbbd360c7ba0eeaf6e755ea3f733e11cf5837b93
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>