rril: Fix issue in recovering signal during stress test
BZ: 95531

When modem is not available, SIM Status is not
updated properly. This results in exception
when framework handles the response for
GET_SIM_STATUS. Also, m_bRadioRequestPending is
set to TRUE after adding the command to command
queue. This may result in wrong state in rapid ril
if the request is handled quickly.

With this patch, SIM state is updated correctly
and m_bRadioRequestPending is set to TRUE
before the command is added to the command queue.
Also, entry logs in functions IsRequestAllowedInRadioOff
and IsRequestAllowedInSpoofState is removed as
this are not needed.

Change-Id: Ic2d001fbd7e839bfd236a32a5895727fe4e37c2e
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/98613
Reviewed-by: D'alton, Alexandre <alexandre.d'alton@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Roques, CyrilX <cyrilx.roques@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>