rril: Fix network loss due to port opening failure
BZ: 96184

On opening of com port failure, modem recovery
is requested in multiple places and also CTE
instance, modem power on, system initialization
events etc are deleted. This results in system
not brought back again properly when the modem
is recovered properly.

With this patch, following changes are done
- Upon com port opening failure, set init flag
as failure and send modem recovery request to MMGR.
- Allow RIL_REQUEST_RADIO_POWER on all
MMGR event
- Remove unused OpenChannelPorts function

Change-Id: I00db28814c564e3a9512542e188677987b65a9b4
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/98986
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>